cart

+++ Entities/Entities/Voucher.cs
@@ -5,9 +5,9 @@
         public Guid Id { get; set; }
         public string? Code { get; set; }
         public int? Percentage { get; set; }
-        public double? MaxDiscount { get; set; }
+        public decimal? MaxDiscount { get; set; }
         public DateOnly? ExpiryDate { get; set; }
-        public double? MinOrderValue { get; set; }
+        public decimal? MinOrderValue { get; set; }
         public int? Quantity { get; set; }
         public ICollection<Order>? Orders { get; set; }
     }
